chore: reorganize structure pages and imports
