+++ src/pages/aurora-site/community/people.jsx
