+++ src/pages/aurora-site/learning/learning-content.jsx
