+++ src/pages/aurora-site/learning/vocabulary-content.jsx
