+++ src/pages/aurora-site/practice-system.jsx
