+++ src/pages/games/game-panel.jsx
