+++ src/pages/games/story-game/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import GameCompleteStory from '../components/games/story-game/game-complete-story';
+import GameCompleteStory from '../../../components/games/story-game/game-complete-story';
 
 const StoryGame = () => {
   const storyText =
